added ability to change colour on different parts
