+++ components/Hoodie/Hoodie.module.css
@@ -1,0 +1,5 @@
+.container {
+  position: relative;
+  width: 774px;
+  height: 700px;
+}
