+++ components/Hoodie/index.js
@@ -1,0 +1,1 @@
+export { Hoodie } from './Hoodie';
